Refactor workload installer responsibilities
Extract package source, inspection, RID selection, and deployment services, and split installer coverage into focused test suites.

Copilot-Session: cee4c3b1-da59-4f88-aef9-38a1cbdd26dc
