+++ src/Func/Workloads/WorkloadPackageTypes.cs
@@ -1,0 +1,10 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace Azure.Functions.Cli.Workloads;
+
+internal static class WorkloadPackageTypes
+{
+    public const string Workload = "FuncCliWorkload";
+    public const string RuntimeIdentifierPackage = "FuncCliWorkloadRidPackage";
+}
